+++ flutter_chaos_toolkit/lib/flutter_chaos_toolkit.dart
@@ -1,5 +1,8 @@
 export 'src/chaos_toolkit.dart';
 export 'src/core/enums/chaos_position.dart';
 export 'src/core/models/chaos_config.dart';
+export 'src/network/chaos_http_client.dart';
+export 'src/network/network_profile.dart';
+export 'src/network/network_throttler.dart';
 export 'src/overlay/chaos_overlay.dart';
 export 'src/overlay/overlay_manager.dart';